--- src/client/common/components/base-network-view/components/index.js
@@ -1,5 +1,0 @@
-const Network = require('./network');
-const Menu = require('./main-menu');
-const Sidebar = require('./sidebar');
-
-module.exports = {Network, Menu, Sidebar};
