+++ src/client/common/components/base-network-view/config/index.js
@@ -1,0 +1,3 @@
+const toolbarButtons = require('./buttons');
+
+module.exports = toolbarButtons;
