+++ src/client/common/components/base-network-view/config/menus/index.js
@@ -1,0 +1,5 @@
+const FileDownloadMenu = require('./file-download');
+const NetworkInfoMenu = require('./network-info');
+
+
+module.exports = {FileDownloadMenu, NetworkInfoMenu};
