+++ src/client/common/components/base-network-view/config/menus/network-info.js
@@ -11,7 +11,7 @@ class NetworkInfoMenu extends React.Component {
       h('p', 'Additional information about the network is normally found here, but we couldn\'t find any for this one.')
     ]);
 
-    const comments = props.comments.map(comment => {
+    const comments = props.networkMetadata.comments.map(comment => {
       return h('div', [h('p', comment.replace(/<p>/g, ' '))]);
     });
