+++ src/styles/features/view/menuBar.css
@@ -55,14 +55,6 @@
     & span {
       cursor: pointer;
     }
-
-    & a {
-      color: var(--link);
-
-      &:visited {
-        color: var(--link-visited);
-      }
-    }
   }
 }
